LayoutTests/fast/scrolling/latching/latching-stuck-to-main-page-expected.txt


Tests that scrolling doesn't get wrongly latched to the main page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking wheel event over the select
PASS selectScrollCount > 0 is true
PASS pageScrollCount == 0 is true
Triggering main page latch; scrolling down
Triggering main page latch; scrolling up
PASS selectScrollCount == 0 is true
PASS pageScrollCount > 0 is true
PASS document.scrollingElement.scrollTop == 0 is true
Checking wheel event over the select again
PASS selectScrollCount > 0 is true
PASS pageScrollCount == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
